change a parameter type for DataBlock.create(size) from int -> long

+++ src/main/java/net/quepierts/animata4j/core/struct/data/DataBlock.java
@@ -13,7 +13,7 @@ public interface DataBlock {
         return MemorySegmentDataBlock.create();
     }
 
-    static MemorySegmentDataBlock create(int size) {
+    static MemorySegmentDataBlock create(long size) {
         return MemorySegmentDataBlock.create(size);
     }
